#10147: Fixed INT32 dtype

+++ tests/tt_eager/python_api_testing/sweep_tests/generation_funcs.py
@@ -2215,7 +2215,7 @@ def gen_bitwise_and_args(
     on_device,
     low=-1,
     high=10,
-    dtype=torch.bfloat16,
+    dtype=torch.int32,
     do_sanitize_args=True,
     coregrid=[],
 ):
